Reorganizado método de execução de pesquisa personalizada
