+++ implementation/solution/src/main/webapp/modules/map/controller/map-controller.js
@@ -2765,7 +2765,7 @@ function MapController($scope, $injector, $log, $state, $timeout, $modal, $locat
             $scope.currentCustomSearch.layerFields[field].value = $("#item_" + field).val();
         }
 
-        markerService.listMarkerByLayerFilters($scope.currentCustomSearch.layer.id, $scope.currentCustomSearch.layerFields, {
+        customSearchService.listMarkerByLayerFilters($scope.currentCustomSearch.layer.id, $scope.currentCustomSearch.layerFields, {
           callback: function (results) {
 
             $scope.markersByLayer = results;
